Bug fix by Long and Juba for spinorbitperatom option:
This happens when you run without SOC using the new solver (non-colinear states without SOC or simply switch off SOC). In this case, you have to read in the file called "kkrflex_spinorbitatom" in which you define which atom you put SOC (1) or not (0). At the moment, the read in values are valid only at the first iteration and after that, from second iteration on, it is set back to the default value of 1 meaning SOC everywhere.