Introduced the input parameters to control the constraint bfields.
The parameters are taken from Sascha Brinkers presentation and should
match what is implemented in KKRhost.
The names are kept as close as possible to the inputcard keywords in
KKRhost, thus they might differ from what they are called in the
KKRhost code.

Added the parameters to datastructures/InputParamsNew.txt and
generated a new datastructures/InputParams_mod.F90 using
the generate.sh skript.