Added bfield to potential where it should be needed.
Calls to add_bfield() whereever the spin-orbit hamiltonian is added to the potential.
Adjusted arguments to the surrounding subroutines and dependencies in the Makefile.

I had to move two subroutines from NonCollinearMagnetism_mod.F90 to a separate
module in NonCollinearMagnetism_Helpers_mod.F90 to avoid circular dependencies.
(They did not change apart from that move.)