Implemented torque.
The subroutine calc_torque() in bfield/torque.f90 calculates the
magnetic torque and iterates the selfconsistency loop for the
constraint magnetic fields, based on torque or on the fields alone.

The routine is called from rhovalnew() as in KKRhost. Adapted the
inputs and included the call. Adapted wrappers_mod, where this
routine is called.

I had to move another subroutine from NonCollinearMagnetism_mod.F90
to the helpers module to avoid circular dependencies. Did not change
the routine otherwise.

Included the new dependencies in the Makefile.

Compiles and did not crash in a short test run,
but not completely tested.