Remove fix angle modes 4 and 5.
Mode 4 was supposed to fix the direction using constraint magnetic
fields but allowing the moment to change direction as far as the
constraint field does not fullfill this yet. While this is in principle
possible, the current implementation was wrong, as the goal angles
are not stored separately, so a changed moment would also change the
goal.
This mode could be added in the future, but as it converges slower
than mode 2 and to the same result, there is not much reason for that.

Mode 5 was supposed to fix the direction by cancelling the torque,
but allowing the moment to move. This does not depend at all on the
given goal directions and does thus not make any sense.