Started working on output for constraint magnetic fields.
I want to incorporate this in the routines where the nonco_angle_out.dat
is written, which is done in a somewhat complicated way. In
ProcessKKRresults (which is called for inMasterGroup, so once per
atom), calculateDensities is called. This subroutine calls
writeResults1File, in which some results are written to a binary file
at an atom-specific position.

As a first step, I moved the call to this subroutine out of
calculateDensities to ProcessKKRresults.

Compiles, does not crash, and gave the same bin.results1 file in
a simple test run I made. Together with what I read in the code
I'm quite confident it did not change anything.