Changed the format of bin.results1 in order to allow including more information.
Instead of one record per atom (which would get to complicated with
more output settings), write several records per atom.
For more details, see comments in the code.

In my calculation, the file got larger by about a factor 2, but as
it is one of the smaller output files this should be fine. In the
long term, all of this will hopefully replaced by MPI communication
anyway...