Fixing a but I introduced for collinear calculations.
As some calculations use the muffin tin index (of the new mesh),
this was calculated in calculateDensities in ProcessKKRresults_mod,
so before checking if a nonco calculation is done. However, in a
collinear calculation, the pointer to the new mesh is not associated,
which gave an error in that case.