Information on convergence of constraint angles only over angles that really are constraint.
The loop calculates the difference in angles before and after they are fixed.
This should always be zero for non-constraint moments (apart from a small deviation
that we not fully understand right now) and converge to zero for constraint moments
(at least approximately, with the torque-based calculation there will be some remaining
difference). For angles constraint only with the old method however, this will converge
to a non-zero value that is not of interest and hides the information on the constraint
moments. Thus, after this commit, they are skipped in the calculation of what is called
"Largest change of angle for constraint moment".