Changed definition of torque (by a factor of the magnetic moment size).
The magnetic torque can be defined either as the derivative of the
total energy with respect to the magnetic moment direction
(T = dE/de) or as the perpendicular part of the derivative of the
total energy with respect to the magnetic moment (T = P dE/dm).
As de = 1/|m| P dm, the two definitions have the same directions
and differ by the factor 1/|m|.

So far, the first definition was used.
However, we decided to switch to the second, as it is easier comparable
to the constraint magnetic field (same units) and thus also makes it
easier to use the torque in scripts or as an initial guess for constraint
magnetic fields.

!! Note that his commit might BREAK COMPATIBILITY with previus scripts !!